+++ packages/rs-dpp/src/address_funds/mod.rs
@@ -1,9 +1,13 @@
 pub mod fee_strategy;
+#[cfg(feature = "shielded-tx")]
+mod orchard_address;
 mod platform_address;
 mod witness;
 mod witness_verification_operations;
 
 pub use fee_strategy::*;
+#[cfg(feature = "shielded-tx")]
+pub use orchard_address::*;
 pub use platform_address::*;
 pub use witness::*;
 pub use witness_verification_operations::*;
