+++ packages/rs-dpp/src/serialization/mod.rs
@@ -1,2 +1,4 @@
+#[cfg(feature = "state-transition-serde-conversion")]
+pub(crate) mod serde_bytes_64;
 pub(crate) mod serialization_traits;
 pub use serialization_traits::*;
