+++ packages/rs-dpp/src/state_transition/state_transitions/shielded/shield_from_asset_lock_transition/v0/mod.rs
@@ -45,7 +45,7 @@ pub struct ShieldFromAssetLockTransitionV0 {
     /// RedPallas binding signature
     #[cfg_attr(
         feature = "state-transition-serde-conversion",
-        serde(with = "crate::shielded::serde_bytes_64")
+        serde(with = "crate::serialization::serde_bytes_64")
     )]
     pub binding_signature: [u8; 64],
     /// ECDSA signature over the signable bytes (excluded from sig hash)
