+++ packages/rs-dpp/src/state_transition/state_transitions/shielded/shield_transition/v0/mod.rs
@@ -49,7 +49,7 @@ pub struct ShieldTransitionV0 {
     /// RedPallas binding signature
     #[cfg_attr(
         feature = "state-transition-serde-conversion",
-        serde(with = "crate::shielded::serde_bytes_64")
+        serde(with = "crate::serialization::serde_bytes_64")
     )]
     pub binding_signature: [u8; 64],
     /// Fee payment strategy
