+++ packages/rs-dpp/src/state_transition/state_transitions/shielded/unshield_transition/v0/mod.rs
@@ -42,7 +42,7 @@ pub struct UnshieldTransitionV0 {
     /// RedPallas binding signature
     #[cfg_attr(
         feature = "state-transition-serde-conversion",
-        serde(with = "crate::shielded::serde_bytes_64")
+        serde(with = "crate::serialization::serde_bytes_64")
     )]
     pub binding_signature: [u8; 64],
 }
